deleted redundant controllers and added all to rest-folder
