+++ src/main/java/org/newton/webshop/rest/AccountController.java
@@ -1,4 +1,4 @@
-package org.newton.webshop.rest.combined;
+package org.newton.webshop.rest;
 
 import org.newton.webshop.models.dto.creation.AccountCreationDto;
 import org.newton.webshop.models.dto.response.AccountDto;
