+++ src/main/java/org/newton/webshop/rest/CustomerController.java
@@ -1,7 +1,7 @@
 package org.newton.webshop.rest;
 
 import org.newton.webshop.models.entities.Customer;
-import org.newton.webshop.services.CustomerService;
+import org.newton.webshop.services.shared.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
