+++ src/main/java/org/newton/webshop/rest/ShoppingController.java
@@ -1,4 +1,4 @@
-package org.newton.webshop.rest.combined;
+package org.newton.webshop.rest;
 
 import org.newton.webshop.models.dto.creation.CustomerCreationDto;
 import org.newton.webshop.models.dto.creation.ItemCreationDto;
