+++ src/main/java/org/newton/webshop/services/StaffService.java
@@ -7,6 +7,8 @@ import org.newton.webshop.models.entities.Address;
 import org.newton.webshop.models.entities.Employee;
 import org.newton.webshop.models.entities.Role;
 import org.newton.webshop.repositories.EmployeeRepository;
+import org.newton.webshop.services.shared.EmployeeService;
+import org.newton.webshop.services.shared.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
