+++ src/main/java/org/newton/webshop/services/shared/CategoryService.java
@@ -1,4 +1,4 @@
-package org.newton.webshop.services;
+package org.newton.webshop.services.shared;
 
 import org.newton.webshop.models.entities.Category;
 import org.newton.webshop.repositories.CategoryRepository;
