+++ src/main/java/org/newton/webshop/services/shared/ProductService.java
@@ -1,4 +1,4 @@
-package org.newton.webshop.services;
+package org.newton.webshop.services.shared;
 
 import org.newton.webshop.models.entities.Product;
 import org.newton.webshop.repositories.ProductRepository;
